requirements.txt
alembic==1.13.1
asgiref==3.7.2
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dill==0.3.8
distlib==0.3.8
Django==5.0
django-environ==0.11.2
django_injector==0.3.1
djangorestframework==3.15.1
filelock==3.13.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
grpcio==1.63.0
grpcio-tools==1.63.0
idna==3.6
injector==0.21.0
itsdangerous==2.2.0
Jinja2==3.1.4
libretranslatepy==2.1.1
lxml==5.1.0
Mako==1.3.3
MarkupSafe==2.1.5
multiprocess==0.70.16
mysqlclient==2.2.4
packaging==24.0
pathos==0.3.2
pillow==10.2.0
platformdirs==4.1.0
pox==0.3.4
ppft==1.7.6.8
protobuf==5.26.1
pycparser==2.21
PyExecJS==1.5.1
PyJWT==2.8.0
PyPDF2==3.0.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
tqdm==4.66.2
translate==3.6.1
translators==5.9.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.25.0
waitress==3.0.0
Werkzeug==3.0.3
